// CoreumV4_Extension_IBC_Hooks/coinflip-fe-next/src/contract/ts/Coinflip.types.ts
/**
* This file was automatically generated by @cosmwasm/ts-codegen@1.11.1.
* DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
* and run the @cosmwasm/ts-codegen generate command to regenerate this file.
*/

export interface InstantiateMsg {}
export type ExecuteMsg = {
  request_random: {
    job_id: string;
  };
} | {
  receive_random: {
    job_id: string;
    randomness: string;
    signature: string;
  };
};
export type QueryMsg = {
  get_all_jobs: {};
} | {
  get_job_by_id: {
    id: string;
  };
};
export interface JobsResponse {
  jobs: JobResponse[];
}
export interface JobResponse {
  id: string;
  randomness: string;
  result: string;
}